requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.8.5
aiogram==3.13.1
beautifulsoup4
icalendar
apscheduler==3.11.1
